Move journal classes into parent package and remove empty package
