+++ hazelcast/src/main/java/com/hazelcast/map/EventJournalMapEvent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map.journal;
+package com.hazelcast.map;
 
 import com.hazelcast.core.EntryEventType;
