+++ hazelcast/src/main/java/com/hazelcast/map/impl/journal/MapEventJournalReadResultSetImpl.java
@@ -18,7 +18,7 @@ package com.hazelcast.map.impl.journal;
 
 import com.hazelcast.map.impl.MapDataSerializerHook;
 import com.hazelcast.internal.serialization.SerializableByConvention;
-import com.hazelcast.map.journal.EventJournalMapEvent;
+import com.hazelcast.map.EventJournalMapEvent;
 import com.hazelcast.projection.Projection;
 import com.hazelcast.ringbuffer.impl.ReadResultSetImpl;
 import com.hazelcast.internal.serialization.SerializationService;
