+++ hazelcast/src/test/java/com/hazelcast/map/impl/journal/EventJournalMapDataStructureAdapter.java
@@ -22,7 +22,7 @@ import com.hazelcast.internal.services.ObjectNamespace;
 import com.hazelcast.journal.EventJournalDataStructureAdapter;
 import com.hazelcast.map.IMap;
 import com.hazelcast.map.impl.MapService;
-import com.hazelcast.map.journal.EventJournalMapEvent;
+import com.hazelcast.map.EventJournalMapEvent;
 import com.hazelcast.ringbuffer.ReadResultSet;
 
 import java.util.Map;
